% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagFullGradientBoostingPrediction.R
\docType{package}
\name{CovCoagFullGradientBoostingPrediction}
\alias{CovCoagFullGradientBoostingPrediction}
\title{CovCoagFullGradientBoostingPrediction}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
